Refine plugin architecture; refactor unit conversion
