+++ specviz/plugins/unit_change/__init__.py
@@ -1,0 +1,1 @@
+from .unit_change_dialog import UnitChangeDialog
